on 06.07.24 @ 15:04

+++ src/righttriangle.c
@@ -32,8 +32,8 @@ int main(int argc, char* argv[]) {
         return EXIT_SUCCESS;
     }
 
-    uint64_t max_perimeter;
-    if (sscanf(argv[1], "%"SCNu64, &max_perimeter) != 1) {
+    uint32_t max_perimeter;
+    if (sscanf(argv[1], "%"SCNu32, &max_perimeter) != 1) {
         puts("  <max_perimeter> must be a positive integer!\n");
         return EXIT_FAILURE;
     }
@@ -52,7 +52,7 @@ int main(int argc, char* argv[]) {
         PEEK_STACK(RightTriangle const* const t, stack)
         PUSH_STACK_N(RightTriangle, next, stack)
         next_rtri(next, t);
-    } while (perimeter_tri(next) <= max_perimeter);
+    } while (perimeter_tri(next) <= (uint64_t)max_perimeter);
     POP_STACK(stack)
 
     for (uint32_t j = stack_size - 1; j != UINT32_MAX; j--) {
@@ -66,7 +66,7 @@ int main(int argc, char* argv[]) {
         do {
             PUSH_STACK_N(RightTriangle, next, stack)
             nextMult_rtri(next, base, k++);
-        } while (perimeter_tri(next) <= max_perimeter);
+        } while (perimeter_tri(next) <= (uint64_t)max_perimeter);
         POP_STACK(stack)
     }
